+++ rs/cli/src/commands/api_boundary_nodes/add.rs
@@ -2,7 +2,8 @@ use clap::Args;
 use ic_types::PrincipalId;
 
 use crate::{
-    commands::{AuthRequirement, ExecutableCommand},
+    auth::AuthRequirement,
+    ctx::exe::ExecutableCommand,
     forum::ForumPostKind,
     ic_admin::{self},
     submitter::{SubmissionParameters, Submitter},
