+++ rs/cli/src/commands/governance/propose/mod.rs
@@ -1,4 +1,3 @@
-use super::{AuthRequirement, ExecutableCommand};
 use clap::Args;
 mod motion;
