+++ rs/cli/src/commands/mod.rs
@@ -1,15 +1,13 @@
 use crate::commands::subnet::Subnet;
+use crate::ctx::exe::impl_executable_command_for_enums;
 use api_boundary_nodes::ApiBoundaryNodes;
-use clap::Args as ClapArgs;
 use clap::Parser;
-use clap_num::maybe_hex;
 use completions::Completions;
 use der_to_principal::DerToPrincipal;
 use firewall::Firewall;
 use get::Get;
 use governance::Governance;
 use hostos::HostOs;
-use ic_canisters::parallel_hardware_identity::KeyIdVec;
 use network::Network;
 use neuron::Neuron;
 use node_metrics::NodeMetrics;
@@ -47,85 +45,7 @@ pub(crate) mod update_unassigned_nodes;
 pub mod upgrade;
 pub(crate) mod version;
 pub(crate) mod vote;
-
-/// HSM authentication parameters
-#[derive(ClapArgs, Debug, Clone)]
-pub(crate) struct HsmParams {
-    /// Slot that HSM key uses, can be read with pkcs11-tool
-    #[clap(required = false,
-        conflicts_with = "private_key_pem",
-        long, value_parser=maybe_hex::<u64>, global = true, env = "HSM_SLOT")]
-    pub(crate) hsm_slot: Option<u64>,
-
-    /// HSM Key ID, can be read with pkcs11-tool
-    #[clap(required = false, conflicts_with = "private_key_pem", long, value_parser=maybe_hex::<u8>, global = true, env = "HSM_KEY_ID")]
-    pub(crate) hsm_key_id: Option<KeyIdVec>,
-}
-
-/// HSM authentication arguments
-/// These comprise an optional PIN and optional parameters.
-/// The PIN is used during autodetection if the optional
-/// parameters are missing.
-#[derive(ClapArgs, Debug, Clone)]
-pub(crate) struct HsmOpts {
-    /// Pin for the HSM key used for submitting proposals
-    // Must be present if slot and key are specified.
-    #[clap(
-        required = false,
-        alias = "hsm-pim",
-        conflicts_with = "private_key_pem",
-        long,
-        global = true,
-        hide_env_values = true,
-        env = "HSM_PIN"
-    )]
-    pub(crate) hsm_pin: Option<String>,
-    #[clap(flatten)]
-    pub(crate) hsm_params: HsmParams,
-}
-
-// The following should ideally be defined in terms of an Enum
-// as there is no conceivable scenario in which both a PEM file
-// and a set of HSM options can be used by the program.
-// Sadly, until ticket
-//   https://github.com/clap-rs/clap/issues/2621
-// is fixed, we cannot do this, and we must use a struct instead.
-// Note that group(multiple = false) has no effect, and therefore
-// we have to use conflicts and requires to specify option deps.
-#[derive(ClapArgs, Debug, Clone)]
-#[group(multiple = false)]
-/// Authentication arguments
-pub struct AuthOpts {
-    /// Path to private key file (in PEM format)
-    #[clap(
-        long,
-        required = false,
-        global = true,
-        conflicts_with_all = ["hsm_pin", "hsm_slot", "hsm_key_id"],
-        env = "PRIVATE_KEY_PEM",
-        visible_aliases = &["pem", "key", "private-key"]
-    )]
-    pub(crate) private_key_pem: Option<String>,
-    #[clap(flatten)]
-    pub(crate) hsm_opts: HsmOpts,
-}
-
-impl TryFrom<String> for AuthOpts {
-    type Error = anyhow::Error;
-
-    fn try_from(value: String) -> std::result::Result<Self, Self::Error> {
-        Ok(AuthOpts {
-            private_key_pem: Some(value),
-            hsm_opts: HsmOpts {
-                hsm_pin: None,
-                hsm_params: HsmParams {
-                    hsm_slot: None,
-                    hsm_key_id: None,
-                },
-            },
-        })
-    }
-}
+use crate::auth::AuthOpts;
 
 #[derive(Parser, Debug)]
 #[clap(version = env!("CARGO_PKG_VERSION"), about, author)]
@@ -182,90 +102,7 @@ The argument is mandatory for testnets, and is optional for mainnet and staging"
     pub cordoned_features_file: Option<String>,
 }
 
-macro_rules! impl_executable_command_for_enums {
-    ($str_name:ident, $($var:ident),*) => {
-        use crate::ctx::DreContext;
-        use clap::{Subcommand, Command};
-
-        #[derive(Subcommand, Debug)]
-        pub enum Subcommands { $(
-            $var($var),
-        )*}
-
-        impl ExecutableCommand for Subcommands {
-            fn require_auth(&self) -> AuthRequirement {
-                match &self {
-                    $(Subcommands::$var(variant) => variant.require_auth(),)*
-                }
-            }
-
-            async fn execute(&self, ctx: DreContext) -> anyhow::Result<()> {
-                match &self {
-                    $(Subcommands::$var(variant) => variant.execute(ctx).await,)*
-                }
-            }
-
-            fn validate(&self, args: &crate::commands::Args, cmd: &mut Command) {
-                match &self {
-                    $(Subcommands::$var(variant) => variant.validate(args, cmd),)*
-                }
-            }
-
-            fn neuron_override(&self) -> Option<crate::auth::Neuron> {
-                match &self {
-                    $(Subcommands::$var(variant) => variant.neuron_override(),)*
-                }
-            }
-        }
-
-        impl ExecutableCommand for $str_name {
-            fn require_auth(&self) -> AuthRequirement {
-                self.subcommands.require_auth()
-            }
-
-            async fn execute(&self, ctx: DreContext) -> anyhow::Result<()> {
-                self.subcommands.execute(ctx).await
-            }
-
-            // Validate the command line arguments. You can return an error with something like:
-            // ```rust
-            // use clap::error::ErrorKind;
-            // if args.neuron_id.is_none() {
-            //    cmd.error(ErrorKind::MissingRequiredArgument, "Neuron ID is required for this command.").exit();
-            // }
-            // ```
-            fn validate(&self, args: &crate::commands::Args, cmd: &mut Command) {
-                self.subcommands.validate(args, cmd)
-            }
-
-            fn neuron_override(&self) -> Option<crate::auth::Neuron> {
-                self.subcommands.neuron_override()
-            }
-        }
-    }
-}
-pub(crate) use impl_executable_command_for_enums;
-
 impl_executable_command_for_enums! { Args, DerToPrincipal, Network, Subnet, Get, Propose, UpdateUnassignedNodes, Version, NodeMetrics, HostOs, Nodes, ApiBoundaryNodes, Vote, Registry, Firewall, Upgrade, Proposals, Completions, Qualify, UpdateAuthorizedSubnets, Neuron, Governance }
-
-pub trait ExecutableCommand {
-    fn require_auth(&self) -> AuthRequirement;
-
-    fn validate(&self, args: &crate::commands::Args, cmd: &mut Command);
-
-    fn execute(&self, ctx: DreContext) -> impl std::future::Future<Output = anyhow::Result<()>>;
-
-    fn neuron_override(&self) -> Option<crate::auth::Neuron> {
-        None
-    }
-}
-
-#[derive(Clone)]
-pub enum AuthRequirement {
-    Anonymous, // for get commands
-    Signer,    // just authentication details used for signing
-    Neuron,    // Signer + neuron_id used for proposals
-}
 
 #[derive(Debug, Display, Clone)]
 pub enum IcAdminVersion {
