+++ rs/cli/src/commands/neuron/mod.rs
@@ -1,6 +1,6 @@
 use clap::Args;
 
-use super::{impl_executable_command_for_enums, AuthRequirement, ExecutableCommand};
+use crate::ctx::exe::impl_executable_command_for_enums;
 use crate::commands::neuron::balance::Balance;
 use crate::commands::neuron::refresh::Refresh;
 use crate::commands::neuron::top_up::TopUp;
