+++ rs/cli/src/commands/proposals/analyze.rs
@@ -4,7 +4,8 @@ use ic_management_types::filter_map_nns_function_proposals;
 use ic_nns_governance::pb::v1::ProposalStatus;
 use registry_canister::mutations::do_change_subnet_membership::ChangeSubnetMembershipPayload;
 
-use crate::commands::{AuthRequirement, ExecutableCommand};
+use crate::auth::AuthRequirement;
+use crate::ctx::exe::ExecutableCommand;
 
 #[derive(Args, Debug)]
 pub struct Analyze {
