+++ rs/cli/src/commands/subnet/deploy.rs
@@ -4,8 +4,10 @@ use ic_types::PrincipalId;
 
 use crate::{
     auth::get_automation_neuron_default_path,
-    commands::{AuthRequirement, ExecutableCommand},
-    forum::ForumPostKind, submitter::{SubmissionParameters, Submitter},
+    auth::AuthRequirement,
+    ctx::exe::ExecutableCommand,
+    forum::ForumPostKind,
+    submitter::{SubmissionParameters, Submitter},
 };
 
 #[derive(Args, Debug)]
