+++ rs/cli/src/commands/subnet/resize.rs
@@ -4,7 +4,7 @@ use ic_management_types::requests::SubnetResizeRequest;
 use ic_types::PrincipalId;
 
 use crate::{
-    commands::{AuthRequirement, ExecutableCommand},
+    auth::{AuthRequirement}, ctx::exe::ExecutableCommand,
     forum::ForumPostKind, submitter::{SubmissionParameters, Submitter},
 };
